+++ workers/tests/unittests/test_helper.py
@@ -14,12 +14,16 @@ from ohsome_quality_analyst.indicators.minimal.indicator import (
 )
 from ohsome_quality_analyst.reports.minimal.report import Minimal as MinimalReport
 from ohsome_quality_analyst.utils.helper import (
+    camel_to_hyphen,
     flatten_dict,
     flatten_sequence,
     get_project_root,
+    hyphen_to_camel,
+    hyphen_to_snake,
     json_serialize,
     loads_geojson,
     name_to_class,
+    snake_to_hyphen,
 )
 
 from .mapping_saturation import fixtures
@@ -41,12 +45,12 @@ class TestHelper(unittest.TestCase):
 
     def test_name_to_class(self):
         self.assertIs(
-            name_to_class(class_type="indicator", name="Minimal"),
+            name_to_class(class_type="indicator", name="minimal"),
             MinimalIndicator,
         )
 
         self.assertIs(
-            name_to_class(class_type="report", name="Minimal"),
+            name_to_class(class_type="report", name="minimal"),
             MinimalReport,
         )
 
@@ -190,6 +194,18 @@ class TestHelper(unittest.TestCase):
         expected = Path(__file__).resolve().parent.parent.parent.resolve()
         result = get_project_root()
         self.assertEqual(expected, result)
+
+    def test_camel_to_hyphen(self):
+        assert camel_to_hyphen("CamelCase") == "camel-case"
+
+    def test_snake_to_hyphen(self):
+        assert snake_to_hyphen("snake_case") == "snake-case"
+
+    def test_hyphen_to_camel(self):
+        assert hyphen_to_camel("hyphen-case") == "HyphenCase"
+
+    def test_hyphen_to_snake(self):
+        assert hyphen_to_snake("hyphen-case") == "hyphen_case"
 
 
 if __name__ == "__main__":